feat: change the way of clearing pipeline
Current version invalidates <prev>_to_<next>_valid signal to invalidate
signals being passed in the pipeline, instead of giving zero values to
signal regs in each pipeline stage, which is adopted in previous version.
It should be noted that the div signals in EXU are given zeros in cases
where dividors cannot give their results in the clock cycle when ex comes.